Refactor: 코드레빗 리뷰사항 반영
